chore: linting issues

+++ src/interfaces/solidly/IUniswapV2Router02.sol
@@ -3,7 +3,6 @@ pragma solidity ^0.8.12;
 
 pragma experimental ABIEncoderV2;
 
-
 interface IUniswapV2Router02 {
     struct route {
         address from;
@@ -20,21 +19,23 @@ interface IUniswapV2Router02 {
     ) external returns (uint256[] memory amounts);
 
     function swapExactTokensForTokensSimple(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address tokenFrom,
         address tokenTo,
         bool stable,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
-    function getAmountsOut(
-        uint amountIn,
-        route[] memory routes) external view returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, route[] memory routes)
+        external
+        view
+        returns (uint256[] memory amounts);
 
     function pairFor(
         address tokenA,
         address tokenB,
-        bool stable) external view returns (address pair);
+        bool stable
+    ) external view returns (address pair);
 }
